+++ pkg/infoschema/builder.go
@@ -36,136 +36,8 @@ import (
 	"github.com/pingcap/tidb/pkg/table"
 	"github.com/pingcap/tidb/pkg/table/tables"
 	"github.com/pingcap/tidb/pkg/util/domainutil"
-	"github.com/pingcap/tidb/pkg/util/logutil"
 	"github.com/pingcap/tidb/pkg/util/sqlexec"
-	"go.uber.org/zap"
 )
-
-type policyGetter struct {
-	is *infoSchema
-}
-
-func (p *policyGetter) GetPolicy(policyID int64) (*model.PolicyInfo, error) {
-	if policy, ok := p.is.PolicyByID(policyID); ok {
-		return policy, nil
-	}
-	return nil, errors.Errorf("Cannot find placement policy with ID: %d", policyID)
-}
-
-type bundleInfoBuilder struct {
-	deltaUpdate bool
-	// tables or partitions that need to update placement bundle
-	updateTables map[int64]any
-	// all tables or partitions referring these policies should update placement bundle
-	updatePolicies map[int64]any
-	// partitions that need to update placement bundle
-	updatePartitions map[int64]any
-}
-
-func (b *bundleInfoBuilder) initBundleInfoBuilder() {
-	b.updateTables = make(map[int64]any)
-	b.updatePartitions = make(map[int64]any)
-	b.updatePolicies = make(map[int64]any)
-}
-
-func (b *bundleInfoBuilder) SetDeltaUpdateBundles() {
-	b.deltaUpdate = true
-}
-
-func (b *bundleInfoBuilder) deleteBundle(is *infoSchema, tblID int64) {
-	delete(is.ruleBundleMap, tblID)
-}
-
-func (b *bundleInfoBuilder) markTableBundleShouldUpdate(tblID int64) {
-	b.updateTables[tblID] = struct{}{}
-}
-
-func (b *bundleInfoBuilder) markPartitionBundleShouldUpdate(partID int64) {
-	b.updatePartitions[partID] = struct{}{}
-}
-
-func (b *bundleInfoBuilder) markBundlesReferPolicyShouldUpdate(policyID int64) {
-	b.updatePolicies[policyID] = struct{}{}
-}
-
-func (b *bundleInfoBuilder) updateInfoSchemaBundles(is *infoSchema) {
-	if b.deltaUpdate {
-		b.completeUpdateTables(is)
-		for tblID := range b.updateTables {
-			b.updateTableBundles(is, tblID)
-		}
-		return
-	}
-
-	// do full update bundles
-	is.ruleBundleMap = make(map[int64]*placement.Bundle)
-	for _, tbls := range is.schemaMap {
-		for _, tbl := range tbls.tables {
-			b.updateTableBundles(is, tbl.Meta().ID)
-		}
-	}
-}
-
-func (b *bundleInfoBuilder) completeUpdateTables(is *infoSchema) {
-	if len(b.updatePolicies) == 0 && len(b.updatePartitions) == 0 {
-		return
-	}
-
-	for _, tbls := range is.schemaMap {
-		for _, tbl := range tbls.tables {
-			tblInfo := tbl.Meta()
-			if tblInfo.PlacementPolicyRef != nil {
-				if _, ok := b.updatePolicies[tblInfo.PlacementPolicyRef.ID]; ok {
-					b.markTableBundleShouldUpdate(tblInfo.ID)
-				}
-			}
-
-			if tblInfo.Partition != nil {
-				for _, par := range tblInfo.Partition.Definitions {
-					if _, ok := b.updatePartitions[par.ID]; ok {
-						b.markTableBundleShouldUpdate(tblInfo.ID)
-					}
-				}
-			}
-		}
-	}
-}
-
-func (b *bundleInfoBuilder) updateTableBundles(is *infoSchema, tableID int64) {
-	tbl, ok := is.TableByID(tableID)
-	if !ok {
-		b.deleteBundle(is, tableID)
-		return
-	}
-
-	getter := &policyGetter{is: is}
-	bundle, err := placement.NewTableBundle(getter, tbl.Meta())
-	if err != nil {
-		logutil.BgLogger().Error("create table bundle failed", zap.Error(err))
-	} else if bundle != nil {
-		is.ruleBundleMap[tableID] = bundle
-	} else {
-		b.deleteBundle(is, tableID)
-	}
-
-	if tbl.Meta().Partition == nil {
-		return
-	}
-
-	for _, par := range tbl.Meta().Partition.Definitions {
-		bundle, err = placement.NewPartitionBundle(getter, par)
-		if err != nil {
-			logutil.BgLogger().Error("create partition bundle failed",
-				zap.Error(err),
-				zap.Int64("partition id", par.ID),
-			)
-		} else if bundle != nil {
-			is.ruleBundleMap[par.ID] = bundle
-		} else {
-			b.deleteBundle(is, par.ID)
-		}
-	}
-}
 
 // Builder builds a new InfoSchema.
 type Builder struct {
@@ -234,24 +106,7 @@ func (b *Builder) ApplyDiff(m *meta.Meta, diff *model.SchemaDiff) ([]int64, erro
 
 func (b *Builder) applyCreateTables(m *meta.Meta, diff *model.SchemaDiff) ([]int64, error) {
 	tblIDs := make([]int64, 0, len(diff.AffectedOpts))
-	if diff.AffectedOpts != nil {
-		for _, opt := range diff.AffectedOpts {
-			affectedDiff := &model.SchemaDiff{
-				Version:     diff.Version,
-				Type:        model.ActionCreateTable,
-				SchemaID:    opt.SchemaID,
-				TableID:     opt.TableID,
-				OldSchemaID: opt.OldSchemaID,
-				OldTableID:  opt.OldTableID,
-			}
-			affectedIDs, err := b.ApplyDiff(m, affectedDiff)
-			if err != nil {
-				return nil, errors.Trace(err)
-			}
-			tblIDs = append(tblIDs, affectedIDs...)
-		}
-	}
-	return tblIDs, nil
+	return b.applyAffectedOpts(m, tblIDs, diff)
 }
 
 func (b *Builder) applyTruncateTableOrPartition(m *meta.Meta, diff *model.SchemaDiff) ([]int64, error) {
@@ -260,6 +115,7 @@ func (b *Builder) applyTruncateTableOrPartition(m *meta.Meta, diff *model.Schema
 		return nil, errors.Trace(err)
 	}
 
+	// bundle ops
 	if diff.Type == model.ActionTruncateTable {
 		b.deleteBundle(b.infoSchema, diff.OldTableID)
 		b.markTableBundleShouldUpdate(diff.TableID)
@@ -284,6 +140,7 @@ func (b *Builder) applyDropTableOrPartition(m *meta.Meta, diff *model.SchemaDiff
 		return nil, errors.Trace(err)
 	}
 
+	// bundle ops
 	b.markTableBundleShouldUpdate(diff.TableID)
 	for _, opt := range diff.AffectedOpts {
 		b.deleteBundle(b.infoSchema, opt.OldTableID)
@@ -296,6 +153,8 @@ func (b *Builder) applyReorganizePartition(m *meta.Meta, diff *model.SchemaDiff)
 	if err != nil {
 		return nil, errors.Trace(err)
 	}
+
+	// bundle ops
 	for _, opt := range diff.AffectedOpts {
 		if opt.OldTableID != 0 {
 			b.deleteBundle(b.infoSchema, opt.OldTableID)
@@ -392,6 +251,7 @@ func (b *Builder) applyRecoverTable(m *meta.Meta, diff *model.SchemaDiff) ([]int
 		return nil, errors.Trace(err)
 	}
 
+	// bundle ops
 	for _, opt := range diff.AffectedOpts {
 		b.markTableBundleShouldUpdate(opt.TableID)
 	}
@@ -432,30 +292,34 @@ func updateAutoIDForExchangePartition(store kv.Storage, ptSchemaID, ptID, ntSche
 	return err
 }
 
+func (b *Builder) applyAffectedOpts(m *meta.Meta, tblIDs []int64, diff *model.SchemaDiff) ([]int64, error) {
+	if diff.AffectedOpts != nil {
+		for _, opt := range diff.AffectedOpts {
+			affectedDiff := &model.SchemaDiff{
+				Version:     diff.Version,
+				Type:        diff.Type,
+				SchemaID:    opt.SchemaID,
+				TableID:     opt.TableID,
+				OldSchemaID: opt.OldSchemaID,
+				OldTableID:  opt.OldTableID,
+			}
+			affectedIDs, err := b.ApplyDiff(m, affectedDiff)
+			if err != nil {
+				return nil, errors.Trace(err)
+			}
+			tblIDs = append(tblIDs, affectedIDs...)
+		}
+	}
+	return tblIDs, nil
+}
+
 func (b *Builder) applyDefaultAction(m *meta.Meta, diff *model.SchemaDiff) ([]int64, error) {
 	tblIDs, err := b.applyTableUpdate(m, diff)
 	if err != nil {
 		return nil, errors.Trace(err)
 	}
 
-	for _, opt := range diff.AffectedOpts {
-		var err error
-		affectedDiff := &model.SchemaDiff{
-			Version:     diff.Version,
-			Type:        diff.Type,
-			SchemaID:    opt.SchemaID,
-			TableID:     opt.TableID,
-			OldSchemaID: opt.OldSchemaID,
-			OldTableID:  opt.OldTableID,
-		}
-		affectedIDs, err := b.ApplyDiff(m, affectedDiff)
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-		tblIDs = append(tblIDs, affectedIDs...)
-	}
-
-	return tblIDs, nil
+	return b.applyAffectedOpts(m, tblIDs, diff)
 }
 
 func (b *Builder) getTableIDs(diff *model.SchemaDiff) (oldTableID, newTableID int64) {
@@ -708,10 +572,15 @@ func (b *Builder) applyModifySchemaCharsetAndCollate(m *meta.Meta, diff *model.S
 			fmt.Sprintf("(Schema ID %d)", diff.SchemaID),
 		)
 	}
-	newDbInfo := b.getSchemaAndCopyIfNecessary(di.Name.L)
-	newDbInfo.Charset = di.Charset
-	newDbInfo.Collate = di.Collate
-	return nil
+	// ywq todo
+	if b.enableV2 {
+		return nil
+	} else {
+		newDbInfo := b.getSchemaAndCopyIfNecessary(di.Name.L)
+		newDbInfo.Charset = di.Charset
+		newDbInfo.Collate = di.Collate
+		return nil
+	}
 }
 
 func (b *Builder) applyModifySchemaDefaultPlacement(m *meta.Meta, diff *model.SchemaDiff) error {
@@ -725,9 +594,14 @@ func (b *Builder) applyModifySchemaDefaultPlacement(m *meta.Meta, diff *model.Sc
 			fmt.Sprintf("(Schema ID %d)", diff.SchemaID),
 		)
 	}
-	newDbInfo := b.getSchemaAndCopyIfNecessary(di.Name.L)
-	newDbInfo.PlacementPolicyRef = di.PlacementPolicyRef
-	return nil
+	// ywq todo
+	if b.enableV2 {
+		return nil
+	} else {
+		newDbInfo := b.getSchemaAndCopyIfNecessary(di.Name.L)
+		newDbInfo.PlacementPolicyRef = di.PlacementPolicyRef
+		return nil
+	}
 }
 
 func (b *Builder) applyDropPolicy(PolicyID int64) []int64 {
@@ -767,6 +641,7 @@ func (b *Builder) applyDropSchema(schemaID int64) []int64 {
 	return tableIDs
 }
 
+// ywq todo
 func (b *Builder) applyRecoverSchema(m *meta.Meta, diff *model.SchemaDiff) ([]int64, error) {
 	if di, ok := b.infoSchema.SchemaByID(diff.SchemaID); ok {
 		return nil, ErrDatabaseExists.GenWithStackByArgs(
@@ -968,12 +843,6 @@ func (b *Builder) applyDropTable(dbInfo *model.DBInfo, tableID int64, affected [
 		}
 	}
 	return affected
-}
-
-// TODO: get rid of this and use infoschemaV2 directly.
-type infoschemaProxy struct {
-	infoschemaV2
-	v1 InfoSchema
 }
 
 // Build builds and returns the built infoschema.
